src/requirements.txt
crewai[bedrock]
crewai-tools
langfuse
python-dotenv
boto3